+++ src/context.h
@@ -24,7 +24,11 @@
 // a round number)
 const uint64_t cyclusYear = 31558200;
 
-const uint64_t kDefaultTimeStepDur = cyclusYear / 12;
+const uint64_t kMonthsPerYear = 12;
+
+const uint64_t cyclusMonth = cyclusYear / kMonthsPerYear;
+
+const uint64_t kDefaultTimeStepDur = cyclusMonth;
 
 const uint64_t kDefaultSeed = 20160212;
 
@@ -249,6 +253,9 @@ class Context {
 
   /// Returns the current simulation timestep.
   virtual int time();
+
+  /// Returns a time shift between the simulation start time and another time stamp
+  int CalcTimeDiff(int year, int month);
 
   /// Adds a package type to a simulation-wide accessible list.
   /// Agents should NOT add their own packages.